LoggerMixin renamed to StatusMixin
