+++ sinner/models/status/Mood.py
